vercelDeployable/api/pyproject.toml
[project]
name = "private-financial-intelligence-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "asyncpg",
    "alembic",
    "prophet",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "bcrypt==4.0.1",
    "pydantic-settings",
    "email-validator",
    "httpx",
    "python-multipart",
    "psycopg2>=2.9.11",
    "google-api-python-client",
    "google-auth",
    "google-auth-oauthlib",
]
